ml-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.5.2
nltk==3.8.1
joblib==1.4.2
gunicorn==21.2.0

